update relative path imports and split files into components and views
