+++ frontend/src/modules/MLStudio/components/MLStudioListItem.js
@@ -15,12 +15,7 @@ import * as FiIcons from 'react-icons/fi';
 import { FiCodesandbox } from 'react-icons/fi';
 import { useNavigate } from 'react-router';
 import { Link as RouterLink } from 'react-router-dom';
-import {
-  IconAvatar,
-  Label,
-  StackStatus,
-  useCardStyle
-} from '../../design';
+import { IconAvatar, Label, StackStatus, useCardStyle } from 'design';
 
 const MLStudioListItem = (props) => {
   const { mlstudiouser } = props;
