+++ frontend/src/modules/MLStudio/views/MLStudioCreateForm.js
@@ -27,14 +27,14 @@ import {
   ChipInput,
   Defaults,
   useSettings
-} from '../../design';
-import { SET_ERROR, useDispatch } from '../../globalErrors';
+} from 'design';
+import { SET_ERROR, useDispatch } from 'globalErrors';
 import {
   createSagemakerStudioUser,
   listEnvironmentGroups,
   listEnvironments,
   useClient
-} from '../../services';
+} from 'services';
 
 const MLStudioCreateForm = (props) => {
   const navigate = useNavigate();
